// src/services/business/schemas/update.schema.ts
export interface BusinessUpdateForm {
  name?: string;
  logoUrl?: string | null;
  street?: string | null;
  number?: string | null;
  complement?: string | null;
  neighborhood?: string | null;
  city?: string | null;
  state?: string | null;
  country?: string | null;
  zipCode?: string | null;
}